+++ vhost_user_block/Cargo.toml
@@ -5,6 +5,7 @@ authors = ["The Cloud Hypervisor Authors"]
 edition = "2018"
 
 [dependencies]
+block_util = { path = "../block_util" }
 clap = { version = "2.33.1", features=["wrap_help"] }
 epoll = ">=4.0.1"
 libc = "0.2.72"
@@ -14,6 +15,5 @@ qcow = { path = "../qcow" }
 vhost_user_backend = { path = "../vhost_user_backend" }
 vhost_rs = { git = "https://github.com/cloud-hypervisor/vhost", branch = "dragonball", package = "vhost", features = ["vhost-user-slave"] }
 virtio-bindings = "0.1.0"
-virtio-devices = { path = "../virtio-devices" }
 vm-memory = "0.2.1"
 vmm-sys-util = ">=0.3.1"
