+++ virtio-devices/Cargo.toml
@@ -12,6 +12,7 @@ mmio_support = []
 [dependencies]
 anyhow = "1.0"
 arc-swap = ">=0.4.4"
+block_util = { path = "../block_util" } 
 byteorder = "1.3.4"
 devices = { path = "../devices" }
 epoll = ">=4.0.1"
