+++ virtio-devices/src/vhost_user/blk.rs
@@ -6,8 +6,8 @@ use super::handler::*;
 use super::vu_common_ctrl::*;
 use super::Error as DeviceError;
 use super::{Error, Result};
-use crate::block::VirtioBlockConfig;
 use crate::VirtioInterrupt;
+use block_util::VirtioBlockConfig;
 use libc::EFD_NONBLOCK;
 use std::cmp;
 use std::io::Write;
